+++ test/bar_signal_strength_indicator_test.dart
@@ -5,53 +5,53 @@ import 'bar_fixture.dart';
 
 void main() {
   group('Bar signal indicators tests', () {
-    testWidgets("Sizing test", (tester) async {
+    testWidgets('Sizing test', (tester) async {
       final widget = createBarSignalStrengthIndicatorSizingTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/bars-sizes.png"),
+        matchesGoldenFile('goldens/bars-sizes.png'),
       );
     });
 
-    testWidgets("Value test", (tester) async {
+    testWidgets('Value test', (tester) async {
       final widget = createBarSignalStrengthIndicatorValueTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/bars-values.png"),
+        matchesGoldenFile('goldens/bars-values.png'),
       );
     });
 
-    testWidgets("Range test", (tester) async {
+    testWidgets('Range test', (tester) async {
       final widget = createBarSignalStrengthIndicatorRangeTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/bars-range.png"),
+        matchesGoldenFile('goldens/bars-range.png'),
       );
     });
 
-    testWidgets("Threshold test", (tester) async {
-      final widget = createBarSignalStrengthIndicatorThresholdTest();
+    testWidgets('Levels test', (tester) async {
+      final widget = createBarSignalStrengthIndicatorLevelsTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/bars-levels.png"),
+        matchesGoldenFile('goldens/bars-levels.png'),
       );
     });
 
-    testWidgets("Bevel test", (tester) async {
+    testWidgets('Bevel test', (tester) async {
       final widget = createBarSignalStrengthIndicatorBevelTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/bars-bevelled.png"),
+        matchesGoldenFile('goldens/bars-bevelled.png'),
       );
     });
   });