+++ test/sector_signal_strength_indicator_test.dart
@@ -6,43 +6,43 @@ import 'sector_fixture.dart';
 
 void main() {
   group('Bar signal indicators tests', () {
-    testWidgets("Sizing test", (tester) async {
+    testWidgets('Sizing test', (tester) async {
       final widget = createSectorSignalStrengthIndicatorSizingTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/sector-sizes.png"),
+        matchesGoldenFile('goldens/sector-sizes.png'),
       );
     });
 
-    testWidgets("Value test", (tester) async {
+    testWidgets('Value test', (tester) async {
       final widget = createSectorSignalStrengthIndicatorValueTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/sector-values.png"),
+        matchesGoldenFile('goldens/sector-values.png'),
       );
     });
 
-    testWidgets("Range test", (tester) async {
+    testWidgets('Range test', (tester) async {
       final widget = createSectorSignalStrengthIndicatorRangeTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/sector-range.png"),
+        matchesGoldenFile('goldens/sector-range.png'),
       );
     });
 
-    testWidgets("Threshold test", (tester) async {
-      final widget = createSectorSignalStrengthIndicatorThresholdTest();
+    testWidgets('Levels test', (tester) async {
+      final widget = createSectorSignalStrengthIndicatorLevelsTest();
 
       await tester.pumpWidget(widget);
       await expectLater(
         find.byType(RepaintBoundary),
-        matchesGoldenFile("goldens/sector-levels.png"),
+        matchesGoldenFile('goldens/sector-levels.png'),
       );
     });
   });